refactor: Remove unused code and use common layout
